+++ src/Derived/DslDerivedFilterRuleMap.php
@@ -5,6 +5,7 @@ namespace HongXunPan\EloquentQueryDsl\Derived;
 use HongXunPan\EloquentQueryDsl\Exception\DslQueryDslDefinitionException;
 use HongXunPan\EloquentQueryDsl\Filter\DslFilterValue;
 use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Database\Eloquent\Model;
 
 /**
  * Query DSL 派生 filter 规则映射。
@@ -31,6 +32,9 @@ class DslDerivedFilterRuleMap implements DslDerivedFilterBehavior
         return $this;
     }
 
+    /**
+     * @param Builder<Model> $query
+     */
     public function apply(Builder $query, DslFilterValue $filterValue): void
     {
         $value = $filterValue->singleValue();