+++ src/Derived/DslNotNullColumnDerivedFilterRule.php
@@ -4,6 +4,7 @@ namespace HongXunPan\EloquentQueryDsl\Derived;
 
 use HongXunPan\EloquentQueryDsl\Exception\DslQueryDslDefinitionException;
 use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Database\Eloquent\Model;
 
 /**
  * Query DSL 派生 filter：字段非空规则。
@@ -27,6 +28,9 @@ class DslNotNullColumnDerivedFilterRule implements DslDerivedFilterRule
         return new self($field);
     }
 
+    /**
+     * @param Builder<Model> $query
+     */
     public function apply(Builder $query): void
     {
         $query->whereNotNull($query->qualifyColumn($this->field));